ubsan: improve testability
Until now, there was no way to verify whether or not UBSAN tests
successfully caught an error. To solve that, we add a new member to the
struct task_struct called ubsan_error to which UBSAN can write the error it
found on the current process, if any. This pointer can then be checked by
automated tests to make sure UBSAN is functioning as intended.

TODO: explain how exactly it is better. "You can access this member
instead of parsing the kernel logs using dmesg"

Signed-off-by: Pedro Sader Azevedo <pedro.saderazevedo@gmail.com>